Edit the smartace service implemention
Now when an ace entry is going to be created, It will be merged with the
previous one with same oid and sid
and same for delete, it will be deducted from the entry.

+++ smart-user-impl/src/main/java/com/smartitengineering/user/security/impl/SmartAceServiceImpl.java
@@ -2,11 +2,11 @@
  * To change this template, choose Tools | Templates
  * and open the template in the editor.
  */
-
 package com.smartitengineering.user.security.impl;
 
 import com.smartitengineering.dao.common.CommonReadDao;
 import com.smartitengineering.dao.common.CommonWriteDao;
+import com.smartitengineering.dao.common.queryparam.FetchMode;
 import com.smartitengineering.dao.common.queryparam.QueryParameter;
 import com.smartitengineering.dao.common.queryparam.QueryParameterFactory;
 import com.smartitengineering.user.domain.UniqueConstrainedField;
@@ -19,10 +19,10 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import org.apache.commons.lang.StringUtils;
 import org.hibernate.StaleStateException;
 import org.hibernate.exception.ConstraintViolationException;
-
 
 /**
  *
@@ -42,7 +42,6 @@ public class SmartAceServiceImpl implements SmartAceService {
         this.smartAclService = smartAclService;
     }
 
-
     public CommonReadDao<SmartAce> getSmartAceReadDao() {
         return smartAceReadDao;
     }
@@ -59,10 +58,19 @@ public class SmartAceServiceImpl implements SmartAceService {
         this.smartAceWriteDao = smartAceWriteDao;
     }
 
-
-
     public void create(SmartAce ace) {
         getSmartAclService().validate(ace.getAcl());
+        SmartAceFilter filter = new SmartAceFilter();
+        filter.setObjectIdentity(ace.getAcl().getObjectIdentity());
+        filter.setSidUsername(ace.getSid().getUsername());
+        List<SmartAce> aces = new ArrayList<SmartAce>();
+        aces = (List<SmartAce>) search(filter);
+        if (aces != null) {
+            SmartAce oldAce = aces.get(0);
+            oldAce.setPermissionMask(ace.getPermissionMask() | oldAce.getPermissionMask());
+            update(oldAce);
+            return;
+        }
         try {
             getSmartAceWriteDao().save(ace);
         } catch (ConstraintViolationException e) {
@@ -75,8 +83,17 @@ public class SmartAceServiceImpl implements SmartAceService {
             throw new RuntimeException(message, e);
         }
     }
+
     public void update(SmartAce ace) {
         getSmartAclService().validate(ace.getAcl());
+        SmartAceFilter filter = new SmartAceFilter();
+        filter.setObjectIdentity(ace.getAcl().getObjectIdentity());
+        filter.setSidUsername(ace.getSid().getUsername());
+        List<SmartAce> aces = new ArrayList<SmartAce>();
+        aces = (List<SmartAce>) search(filter);
+        if (aces != null) {
+            ace.setPermissionMask(aces.get(0).getPermissionMask() | ace.getPermissionMask());
+        }
         try {
             getSmartAceWriteDao().update(ace);
         } catch (ConstraintViolationException e) {
@@ -91,11 +108,32 @@ public class SmartAceServiceImpl implements SmartAceService {
     }
 
     public void delete(SmartAce ace) {
-        try {
-            getSmartAceWriteDao().delete(ace);
-        } catch (RuntimeException e) {
-            String message = ExceptionMessage.CONSTRAINT_VIOLATION_EXCEPTION.name() + "-" + UniqueConstrainedField.ACE;
-            throw new RuntimeException(message, e);
+        SmartAceFilter filter = new SmartAceFilter();
+        filter.setObjectIdentity(ace.getAcl().getObjectIdentity());
+        filter.setSidUsername(ace.getSid().getUsername());
+        List<SmartAce> aces = new ArrayList<SmartAce>();
+        aces = (List<SmartAce>) search(filter);
+        if (aces != null) {
+            SmartAce oldAce = aces.get(0);
+
+            if (oldAce.getPermissionMask() != ace.getPermissionMask()) {
+
+                // to delete permission from a existing ACL entry, we need to do the followings
+                //   1. Firstly, bitwise OR operation of new and old permission mask
+                //   2. Secondly, numerical minus of new permission mask from the result mask
+                //        In expression :   { (old | new) - new }
+
+                oldAce.setPermissionMask((ace.getPermissionMask() | oldAce.getPermissionMask()) - ace.getPermissionMask());
+                update(oldAce);
+                return;
+            }
+
+            try {
+                getSmartAceWriteDao().delete(ace);
+            } catch (RuntimeException e) {
+                String message = ExceptionMessage.CONSTRAINT_VIOLATION_EXCEPTION.name() + "-" + UniqueConstrainedField.ACE;
+                throw new RuntimeException(message, e);
+            }
         }
     }
 
@@ -103,13 +141,15 @@ public class SmartAceServiceImpl implements SmartAceService {
         QueryParameter qp = null;
         List<QueryParameter> queryParameters = new ArrayList<QueryParameter>();
         if (!StringUtils.isEmpty(filter.getObjectIdentity().getOid())) {
-            qp = QueryParameterFactory.getEqualPropertyParam("oid",
-                    filter.getObjectIdentity().getOid());
+            qp = QueryParameterFactory.getNestedParametersParam("acl", FetchMode.DEFAULT,
+                    QueryParameterFactory.getEqualPropertyParam(
+                    "objectIdentity.oid", filter.getObjectIdentity().getOid()));
             queryParameters.add(qp);
         }
         if (!StringUtils.isEmpty(filter.getSidUsername())) {
-            qp = QueryParameterFactory.getEqualPropertyParam("oid",
-                    filter.getSidUsername());
+            qp = QueryParameterFactory.getNestedParametersParam("sid", FetchMode.DEFAULT,
+                    QueryParameterFactory.getEqualPropertyParam("username",
+                    filter.getSidUsername()));
             queryParameters.add(qp);
         }
